src/core/ppt/requirements.txt
# PPT模块依赖包
# 核心依赖
aiohttp>=3.8.0
python-pptx>=0.6.21
jinja2>=3.1.0
PyYAML>=6.0

# 可选依赖（用于增强功能）
# 图像处理
Pillow>=9.0.0  # 用于图像处理

# 测试依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# 类型检查
mypy>=1.0.0

# 代码质量
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0